Fix: Correctly format countdown string
The dateToEndsInXdYhZm function was producing a countdown string with a leading space when one of the time components (days, hours) was zero.

This commit fixes the issue by filtering out empty time components before joining them into a string.

A permanent test case has been added to verify the fix.
